Add files via upload
Algorithmic trading also called as automated trading is the
process of using computers programmed to follow a defined set
of instructions for placing a trade in order to generate profits at a
speed and frequency that is impossible for a human trader. An
important factor affecting the trading decisions is the ability of
predicting stock market movement.